Add addRemove.js and modify index.js
