+++ src/status.js
@@ -1,6 +1,7 @@
-export default function statusUpdate(event) {
-  event.target.nextElementSibling.classList.toggle('completed');
-  const status = JSON.parse(localStorage.getItem('listCollection'));
-  status[event.target.dataset.id].completed = event.target.checked;
-  localStorage.setItem('listCollection', JSON.stringify(status));
+function updateStatus(e, i) {
+  const listSaved = JSON.parse(localStorage.getItem('listSaved'));
+  listSaved[i].completed = !listSaved[i].completed;
+  e.target.checked = listSaved[i].completed;
+  localStorage.setItem('listSaved', JSON.stringify(listSaved));
 }
+export default updateStatus;
