+++ webpack.config.js
@@ -3,7 +3,7 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-
+  mode: 'development',
   entry: './src/index.js',
   output: {
     filename: 'main.js',
